+++ Etomo/src/etomo/ui/swing/NewstackOrBlendmontPanel.java
@@ -38,6 +38,9 @@ import etomo.util.InvalidParameterException;
  * @version $Revision$
  * 
  * <p> $Log$
+ * <p> Revision 1.2  2011/02/22 18:16:52  sueh
+ * <p> bug# 1437 Reformatting.
+ * <p>
  * <p> Revision 1.1  2010/11/13 16:07:34  sueh
  * <p> bug# 1417 Renamed etomo.ui to etomo.ui.swing.
  * <p>
@@ -185,7 +188,7 @@ abstract class NewstackOrBlendmontPanel implements Run3dmodButtonContainer, Expa
     newstackAndBlendmontParamPanel.setFiducialessAlignment(input);
   }
 
-  void setImageRotation(float input) {
+  void setImageRotation(double input) {
     newstackAndBlendmontParamPanel.setImageRotation(input);
   }
